Pitter/src/requirements.txt
Django==2.1.7
python-dotenv==0.10.1
psycopg2-binary==2.8.1
djangorestframework==3.9.2
redis==3.2.1
django_redis==4.10.0
requests==2.21.0
djangorestframework-jwt==1.11.0
celery>=4.2,<4.3
flake8==3.7.7
cryptography==2.6.1
sentry-sdk==0.7.11
